Move marker undo code to seperate file
